+++ app/src/main/res/layout/activity_server_crawling.xml
@@ -20,198 +20,15 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.02" />
 
-    <ImageButton
-        android:id="@+id/Btn_refresh"
-        android:layout_width="48dp"
-        android:layout_height="48dp"
-        android:background="#01DF01"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.95"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.02"
-        app:srcCompat="@android:drawable/ic_popup_sync" />
-
     <WebView
         android:id="@+id/Web_server_data"
         android:layout_width="370dp"
-        android:layout_height="320dp"
-        tools:layout_editor_absoluteX="1dp"
-        tools:layout_editor_absoluteY="389dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintVertical_bias="0.965"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
-
-    <TableLayout
-        android:layout_width="360dp"
-        android:layout_height="324dp"
-        android:gravity="center"
+        android:layout_height="640dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.165">
-
-        <TableRow
-            android:background="@android:color/background_dark"
-            android:gravity="center">
-
-            <TextView
-                android:layout_width="72dp"
-                android:layout_height="32dp"
-                android:layout_margin="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="조\\센서"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="1조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="2조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="3조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="4조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="5조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-        </TableRow>
-
-        <TableRow
-            android:background="@android:color/background_dark"
-            android:gravity="center">
-
-            <TextView
-                android:layout_width="72dp"
-                android:layout_height="32dp"
-                android:layout_margin="1dp"
-                android:background="#C8C8C8"
-                android:gravity="center"
-                android:text="2조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/Text_1jo"
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="@color/white"
-                android:gravity="center"
-                android:hint="1조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/Text_2jo"
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="@color/white"
-                android:gravity="center"
-                android:hint="2조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/Text_3jo"
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="@color/white"
-                android:gravity="center"
-                android:hint="3조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/Text_4jo"
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="@color/white"
-                android:gravity="center"
-                android:hint="4조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/Text_5jo"
-                android:layout_width="56dp"
-                android:layout_height="32dp"
-                android:layout_marginTop="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_marginBottom="1dp"
-                android:background="@color/white"
-                android:gravity="center"
-                android:hint="5조"
-                android:textSize="16sp"
-                android:textStyle="bold" />
-        </TableRow>
-    </TableLayout>
+        app:layout_constraintVertical_bias="0.8" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
